Rename objects to make it clear it's about ccd
